+++ src/main/java/com/github/romualdrousseau/shuju/transforms/NumericScaler.java
@@ -8,7 +8,7 @@ import com.github.romualdrousseau.shuju.features.NumericFeature;
 public class NumericScaler implements ITransform {
     public NumericScaler(DataSummary summary) {
         this.min = summary.min;
-        this.ratio = 1.0 / (summary.max - summary.min);
+        this.ratio = 1.0f / (summary.max - summary.min);
     }
 
     public void apply(IFeature<?> feature, int rowIndex, int colIndex) {
@@ -17,6 +17,6 @@ public class NumericScaler implements ITransform {
         numericFeature.setValue((numericFeature.getValue() - this.min) * this.ratio);
     }
 
-    private double min;
-    private double ratio;
+    private float min;
+    private float ratio;
 }
